fix: 修复缓存config.json错误

+++ packages/uni-uts-v1/src/utils.ts
@@ -699,7 +699,7 @@ export function genConfigJson(
     platform
   )
   const configJsonDestFile = resolve(utsOutputDir, 'config.json')
-  if (fs.existsSync(configJsonDestFile)) {
+  if (configJsonCacheFile && fs.existsSync(configJsonDestFile)) {
     fs.copySync(configJsonDestFile, configJsonCacheFile)
   }
 }
